rsuper_train/requirements.txt
mmcv-full
einops==0.6.0
PyYAML==6.0
scikit-image==0.20.0
scikit-learn==1.2.2
SimpleITK==2.2.1
timm==0.6.12
torch==2.1.0
tqdm==4.65.0
matplotlib==3.8.4
scipy
pandas
tensorboard
monai
batchgenerators
cvxpy
seaborn
nibabel
